# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "The Tattoo API",
    "description": "An API that serves Tattoo, Style, and Color data."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/tattoos/": {
      "get": {
        "tags": [
          "Tattoos"
        ],
        "summary": "Retrieve a list of tattoos from the database.",
        "description": "Description of /tattoos",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/tattoos/{id}": {
      "get": {
        "tags": [
          "Tattoos"
        ],
        "summary": "Retrieve a single tattoo object by ID.",
        "description": "Description of /tattoos/:id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}